+++ index.d.ts
@@ -1,7 +1,0 @@
-/// <reference path="./react.d.ts" />
-/// <reference path="./jquery.d.ts" />
-/// <reference path="./makeGroups.d.ts" />
-
-export * from "./formatchange";
-import x from "./formatchange";
-export default x;